requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
anthropic>=0.40.0
qdrant-client==1.7.3
sentence-transformers==2.3.1
PyMuPDF==1.23.22
pydantic-settings==2.1.0
python-multipart==0.0.9
google-generativeai>=0.3.2
openai>=1.12.0
cohere>=5.0.0
